backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
python-dotenv
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
python-dateutil